% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arandi.R
\name{arandi}
\alias{arandi}
\title{Port of the arandi function from the mcclust package which is no longer maintained.}
\usage{
arandi(cl1, cl2, adjust = TRUE)
}
\description{
Port of the arandi function from the mcclust package which is no longer maintained.
}
\keyword{internal}
