% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbR-data.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{Example, symmetric weighted count matrix}
\format{A 25 by 25 symmetric count matrix.}
\usage{
exampledata
}
\description{
This object contains a simulated 25 by 25 symmetric, weighted count matrix.
}
\keyword{datasets}
