% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb 7.R
\name{perturb}
\alias{perturb}
\title{Perturb networks and evaluate subgroup structures.}
\usage{
perturb(matrix, method)
}
\arguments{
\item{matrix}{A symmetric, weighted matrix or weighted graph}
}
\description{
Perturb networks and evaluate subgroup structures.
}
