% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbR-pkg.R
\name{perturbR-package}
\alias{perturbR-package}
\title{Perturb Networks}
\description{
The perturbR function incrementally perturbs networks 
(using the rewireR function) and compares the resulting rewired networks with the orignal. 
These comparisons aid in understanding the stability of the 
cluster solution.
}
\author{
{KM Gates [aut, cre],\cr
    Zachary Fisher [aut], \cr
    Cara Arizmendi [aut]	
}
Maintainer: KM Gates \email{gateskm@email.unc.edu}
}
\keyword{perturbR}
