% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.dist.R
\name{vi.dist}
\alias{vi.dist}
\title{Port of the vi.dist function from the mcclust package which is no longer maintained.}
\usage{
vi.dist(cl1, cl2, parts = FALSE, base = 2)
}
\description{
Port of the vi.dist function from the mcclust package which is no longer maintained.
}
\keyword{internal}
