% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalClust.R
\name{perturb}
\alias{perturb}
\alias{evalClust}
\title{Perturb networks and evaluate subgroup structures.}
\usage{
evalClust(sym.matrix, plot = TRUE, resolution = 0.01, reps = 100)
}
\arguments{
\item{sym.matrix}{A symmetric, weighted matrix or graph object}

\item{plot}{Logical, defaults to TRUE}

\item{resolution}{The percentage of edges to iteratively alter. One percent is default, increase to go quicker.}

\item{reps}{The number of repititions to do for each level of perturbation. Decrease to make it go quicker.}
}
\description{
Randomly rewires networks in increasing degrees of
   perturbation to evaluate stability of community solutions obtained from Walktrap.
}
\examples{
#Reps should be 100.  Reps set equal to 1, so example can run more quickly.
perturbR(exampledata, plot=TRUE, resolution=0.01, reps=1)
}
\keyword{evalClust}
